+++ src/Components/Pages/FiveDayForecast.tsx
@@ -1,13 +1,14 @@
 import { FC } from 'react';
 import { ThreeHourForecast } from '../Weather/ThreeHourForecast';
-import { List } from '../../Types/ApiResponses/FiveDayForecast';
+import { City, List } from '../../Types/ApiResponses/FiveDayForecast';
 
 interface FiveDayForecastPageProps {
-	currentForecast: List[] | null;
+	currentCity?: City;
+	currentForecast?: List[];
 }
 
 export const FiveDayForecastPage: FC<FiveDayForecastPageProps> = (props) => {
-	const { currentForecast } = props;
+	const { currentCity, currentForecast } = props;
 	return(
 		<div className="five-day-forecast-page">
 			<h1>Five Day Forecast</h1>
@@ -16,11 +17,15 @@ export const FiveDayForecastPage: FC<FiveDayForecastPageProps> = (props) => {
 				const temperatureIcon = threeHourBlock.weather[0].icon;
 				const temperatureValue = threeHourBlock.main.temp;
 				const weatherDescription = threeHourBlock.weather[0].description;
+				const sunrise = currentCity?.sunrise;
+				const sunset = currentCity?.sunset;
 
 				return(
 					<ThreeHourForecast
 						key={threeHourBlock.dt}
 						forecastPeriod={forecastPeriod}
+						sunrise={sunrise}
+						sunset={sunset}
 						temperatureIcon={temperatureIcon}
 						temperatureValue={temperatureValue}
 						weatherDescription={weatherDescription}